ci-operator/jobs/openshift/release: Drop --rehearsal-limit
Folks have been tuning this since b0d60e3273 (Bump the limit of
rehearsed jobs, 2019-08-19, #4789), most recently in 3f0040eca8 (Raise
rehearsal-limit to 45 temporarily, 2020-02-03, #6999).  But with [1],
we no longer need to fuss with this setting in order to see rehearsals
for changes that touch lots of jobs, so let it fall back to
pj-rehearse's default of 15.

[1]: https://github.com/openshift/ci-tools/pull/1315