ci-operator/jobs/openshift/release: Use upgrade-all in chained updates
Using the 'upgrade-all' precedent from cfcd60fed1 (release:
Standardize all ci-chat-bot jobs, 2020-04-27, #8594).  I'm not clear
on why we are joining with a newline instead of '&&'; presumably this
is getting wrapped in a 'set -e' or equivalent.  But I'm sticking with
newline to match precedent.

This increases the risk that we time out these slow jobs (e.g. [1]
took 3h42m), but we really want to exercise tests like
openshift/origin@9f7fe0089d (Add test for scaling machineSets,
2019-04-11, openshift/origin#22564), which is in
openshift/conformance/serial, because machines launch with the born-in
boot images until we get [2].

And in fact, the reason why we didn't have this post-update suite in
4.6 was because of 3bc9d8e6af (stop running e2e tests after three
upgrades because we hit timeouts and lose upgrade signal, 2020-10-05, #12436).
But since 3c915e2f0c (ci-operator/step-registry/openshift/e2e/test:
Add 2h active_deadline_seconds, 2020-10-09, #12647), we no longer have
to worry about getting logs when that step is slow.  So we might not
pass if we're slow, but we'll still get logs to debug why we're slow.

Only for 4.6 and later, because 4.5 is live and if we had problems
there we'd probably have already heard about them from customers.

[1]: https://prow.ci.openshift.org/view/gs/origin-ci-test/logs/release-openshift-origin-installer-e2e-aws-upgrade-4.3-to-4.4-to-4.5-to-4.6-ci/1318709056830967808
[2]: https://github.com/openshift/enhancements/pull/201