WIP: ci-operator/config/openshift/machine-config-operator: Add e2e-upgrade-minor
Some issues only turn up in minor-version bumps [1], and with this
change machine-config developers will be able to comment on a pull
request with:

  /test e2e-upgrade-minor

to launch a job that installs the previous minor's OCP candidate
release and updates to a release built from the in-flight pull
request.

Generated by manually editing ci-operator/config following [2], running:

  $ make update

and editing the configurable always_run and optional properties [3].

WIP because I'm waiting on #12959.  I'll rebase this on top after that lands.

[1]: https://bugzilla.redhat.com/show_bug.cgi?id=1890250#c5
[2]: https://github.com/openshift/ci-docs/blame/047e38285a227888733b1f5a49fd844d047d50f8/content/en/docs/architecture/ci-operator.md#L278-L283
[3]: https://github.com/openshift/ci-tools/blob/d45a03ce3a38d79a1b7bf82f35919505dc3d1c9e/GENERATOR.md#hand-edited-prow-configuration