ci-operator/config/openshift/cincinnati: Drop builder caching for 4.5+
The Cincinnati repository requires a Rust toolchain to build its
images.  There is no official image holding a Rust, so the Cincinnati
repository includes a Dockerfile for creating a Rust-toolchain image
[1].  Building that image takes some time, so instead of building it
on-demand for each PR, our postsubmit jobs promote it to the
cincinnati-ci namespace for use in later presubmits since 05d0f17364
(ci-operator/config/openshift/cincinnati/openshift-cincinnati-master:
use cached build root image, 2020-04-07, #9857).  The benefit is that
those later presubmits can just pull an existing builder image, which
is much faster.  The downsides are:

* The cached builder may be stale, because the postsubmit which built
  it may have run before subsequent RPM or CentOS fixes.

* The cached builder will not reflect changes made to the builder's
  Dockerfile in the pull request which changes that Dockerfile.
  Changes to the builder image will only affect the builder image
  pushed into the cache by a postsubmit job.  This makes it harder to
  review pull requests that change the builder's Dockerfile.

* There may be no cached builder after branch creation, because you
  need a successful postsubmit run to create it.  For example, until a
  branch is advanced via a pull request merging, no postsubmits have
  been run on it, which can fail presubmits on pull requests
  targetting that branch.  Currently, a cached builder exists for
  master and 4.4, but not for the other branches:

    $ for SUFFIX in '' -release-4.{4,5,6,7,8}; do echo "suffix: ${SUFFIX}"; oc image info "registry.svc.ci.openshift.org/cincinnati-ci/cincinnati-build-root${SUFFIX}:buildroot" 2>&1 | head -n2; done
    suffix:
    Name:        registry.svc.ci.openshift.org/cincinnati-ci/cincinnati-build-root:buildroot
    Digest:      sha256:4cdbf6b512be4536a29924993b2fd6172871ff8b0ba976de78afb20ff2a91665
    suffix: -release-4.4
    Name:        registry.svc.ci.openshift.org/cincinnati-ci/cincinnati-build-root-release-4.4:buildroot
    Digest:      sha256:1441c046a0d8d7b75fc75659d3bd12a43205dd6e05c77c0500753bd3d9b22c98
    suffix: -release-4.5
    error: unable to read image registry.svc.ci.openshift.org/cincinnati-ci/cincinnati-build-root-release-4.5:buildroot: name unknown
    suffix: -release-4.6
    error: image does not exist
    suffix: -release-4.7
    error: image does not exist
    suffix: -release-4.8
    error: unable to read image registry.svc.ci.openshift.org/cincinnati-ci/cincinnati-build-root-release-4.8:buildroot: name unknown

To break out of the presubmit/postsubmit dependency cycle and avoid
the other two cached-builder issues, this commit removes the cached
image consumer and compiles a fresh builder image for each presubmit
run in 4.5 through 4.8 jobs.  The slower presubmits aren't as much of
a problem, because backports into those branches shouldn't be all that
common.  I am keeping the postsubmit promotions, so if we end up
deciding that the slower presubmits are tedious, we can revert this
commit and have the dependency cycle already broken with the cached
images in place for future consumption.

I'm a bit surprised that there is no slot to override the 'from' under
project_image, like there is for the 'images' entries, but there isn't
[2,3] (it is unique to ProjectDirectoryImageBuildStepConfiguration
[4]).  So I guess these builds just use the CentOS base we configure
in the Dockerfile itself, without getting the CI-registry-hosted image
stream plugged in.

[1]: https://github.com/openshift/cincinnati/blob/934dc4ae68a7d0085f1f442589fa2c39600b1905/dist/Dockerfile.build/Dockerfile
[2]: https://github.com/openshift/ci-tools/blob/f1f806730dbec7c9efeb7b4f79f6f975d604f3fe/pkg/api/types.go#L322
[3]: https://github.com/openshift/ci-tools/blob/f1f806730dbec7c9efeb7b4f79f6f975d604f3fe/pkg/api/types.go#L1113-L1127
[4]: https://github.com/openshift/ci-tools/blob/f1f806730dbec7c9efeb7b4f79f6f975d604f3fe/pkg/api/types.go#L1102