Use upstream.dev Dockerfile to build PAO 4.6 bundle
The PR should fix the problem when the bundle image was created with manifests without a real image.
It included the REPLACE_IMAGE instead of the real image.

So now we use Dockerfile.bundle.ci for the CI operator flow and the Dockerfile.upstream.dev to build
bundle for the upstream usage as to use it under the cnf-features-deploy repository.

Signed-off-by: Artyom Lukianov <alukiano@redhat.com>