ci-operator/config/openshift/cluster-baremetal-operator: Add platform-agnostic e2e
This will protect us from "operator works on installer-provisioned
metal, but chokes on non-metal platforms" [1].

Using a platform-agnostic job, as described in
ci-operator/platform-balance, so CI admins can shift the backing
implementation around as needed.  Pretty sure they won't switch it to
metal-IPI, but there's no syntax for explicitly declaring "but not..."
today, so that is a small risk.

Generated by editing ci-operator/config and then running:

  $ make update

to update ci-operator/jobs.

[1]: https://github.com/openshift/cluster-baremetal-operator/pull/60