oVirt: take resources value from ovirt secret
Because till now the resorces where hard coded into the script
whenever we wanted to change them we needed to send a PR a process that take some time and should be in our full control.
Because we still blocked on porting to workflows we changed the lease pod logic to take the parameters for the mem and cpu from a secret in ovirt Namespace.
This will help make the CI easier to maintain and configure

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>