Remove all usage of openshift_installer_console
These jobs are actually using the
`openshift_installer_custom_test_image` template (see the CM that is
mounted for the `job-definition` volume), so make the inline ci-op
config consistent with that. These stanzas are only actually consumed by
prowgen (and not by ci-operator), so they only need to be parseable by
ci-operator. Changing them here will allow us to remove the consuming
code in prowgen and related types.