Remove image build, inject oc, call script directly
As described in https://docs.ci.openshift.org/docs/architecture/step-registry/#injecting-the-oc-cli
we can just use the oc cli injection that provides us with the oc binary and
get rid of the image build.

Furthermore we just call the test script directly now. Unfortunately we
have to declare a couple of env vars for now, this needs to get moved
into the kubevirt ci script.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>