ci-operator/config/openshift/cluster-version-operator: Multi-step operator tests
The outgoing integration tests have used a master-sidecar since they
landed in d1af815627 (Integration tests for registry, replacing
existing, 2018-05-23, #888).  This recently broke, with jobs like [1]
failing with:

  Waiting for kube-apiserver to come up
  !!! [1118 23:29:05] Timed out waiting for kube-apiserver:  to answer at https://10.130.11.147:8443/healthz; tried 60 waiting 1 between each

With the Kubernetes API-server dying with:

  $ curl -s https://gcsweb-ci.apps.ci.l2s4.p1.openshiftapps.com/gcs/origin-ci-test/pr-logs/pull/opensift_cluster-version-operator/443/pull-ci-openshift-cluster-version-operator-master-integration/1329201271097266176/artifacts/integration/logs/kube-apiserver.log | tail -n2
  I1118 23:28:05.064590     308 flags.go:59] FLAG: --watch-cache-sizes="[]"
  Error: error creating self-signed certificates: mkdir /var/run/kubernetes: permission denied

This commit pivots to the off-the-shelf openshift-e2e-gcp workflow,
substituting in a custom 'test' step to run the existing integration
tests.

Porting back to 4.4.  Older releases are end-of-life [2], but if we
need to revive them for some reason we can port this back further
then.

[1]: https://prow.ci.openshift.org/view/gs/origin-ci-test/pr-logs/pull/openshift_cluster-version-operator/443/pull-ci-openshift-cluster-version-operator-master-integration/1329201271097266176
[2]: https://access.redhat.com/support/policy/updates/openshift#dates