ci-operator: Remove EOL branches for the metering-related components
Removes the ci-operator/config for the 4.1/4.2/4.3 branches, which are marked as EOL as of the 4.6 GA release.

The ci-operator/jobs were removed after running `$ make ci-operator-config && make jobs` locally where the `make jobs` target is responsible for propagating changes made to the config file to the corresponding job for each release branch affected.