baremetalds: update url for oc util
For whatever reason, after this URL has been working for ages,
https://mirror.openshift.com/pub/openshift-v4/clients/oc/4.4/linux/oc.tar.gz
just no longer exists.

A lot of jobs in this repo rely on it.  For baremetal, we rely on it to
bootstrap extracting the client from the release payload. This PR
updates the URL to grab the latest stable release from
mirror.openshift.com. I don't know if it's temporary or not, but it
happened sometime near the end of business today.

It's blocking a ton of work.