Bump packet boskos leases limit
We were at the limit of 20 hosts nearly all day, delaying some work as
we get close to FF. Multiple teams are relying on the baremetal IPI
tests, for example the assisteed installer team also relies on the
Packet cluster type.  We peak at 10 leases during an average day but we
want to be able to burst a bit higher than 20 if needed.