Promote assisted-test-infra image to openshift-kni namespace
This change is required to stop promoting the asssited-test-infra test image in the ocp/4.x image stream, since this forces the assisted-test-infra repo to fall within the OpenShift CI centralized release branch management, and we'd like to avoid that.
The fix then promotes the assisted-test-infra image into the openshift-kni/assisted-test image stream.