build01: Add a single m5.metal instance
To fix https://github.com/openshift/os/issues/417

This will allow the CoreOS group and others (openshift libvirt testing)
to directly run KVM virtual machines as pods inside the cluster.

The pattern here is much like what is necessary for OpenShift virtualization
(or VMWare) in AWS; we just want to run pods that run ephemeral VMs
as opposed to more "pet" VMs.

xref https://issues.redhat.com/browse/DPTP-1697

For now...I didn't add a taint because it will allow us to more easily
test things, plus I think it's better to allow a mixed workload
because otherwise the hardware will be going to waste a fair bit of the
time probably.

If we keep having problems with workloads-requiring-metal not scheduling
quickly, we could bump up to e.g. `maxReplicas: 2` or so.