WIP: ci-operator/step-registry/ipi/deprovision/pruner: Port from ci-tools
The test-platform folks currently maintain a pruner that covers both
AWS and GCP [1].  However, when we fail to remove clusters, issues
tend to be:

* Installer bugs, because the installer is failing to discover and
  delete some blocking resource (e.g. [2]).
* Provider bugs (e.g. AWS flaking out on eventual-consistency [3]).
* Component/test bugs (e.g. someone creating a resource that the
  installer doesn't know how to find and reap).

Moving the reaper into the step registry makes it easier to maintain
(just the script, we can drop the old ipi-deprovision image), and
having platform-specific periodics allows the test-platform folks to
route job-failed alerts to the appropriate platform-CI admins (where
the old multi-platform reaper needed manual inspection to determine
which platform's reaping was failing before you knew who to ping about
the failure).

The 'cron' entry runs this at X:51, every even hour.

In openshift-release-master__ocp-4.8.yaml, because we need releases:
entries to avoid [4]:

  2020/12/09 23:21:30 step <*steps.leaseStep> is missing dependencies: <&api.internalImageStreamTagLink{name:"release", tag:"latest", unsatisfiableError:""}>
  2020/12/09 23:21:30 Ran for 0s
  error: some steps failed:
    * could not print execution order: could not sort nodes: steps are missing dependencies

It will probably need to get dragged along as we fork releases.

[1]: https://github.com/openshift/ci-tools/tree/f09874d38e543b3fa09bc8ab8a8d9ffaace35015/cmd/ipi-deprovision
[2]: https://github.com/openshift/installer/pull/4440
[3]: https://bugzilla.redhat.com/show_bug.cgi?id=1760616
[4]: https://github.com/openshift/release/pull/14226#discussion_r539727236