app.ci: add permissions to OKD admins group
In order to mirror OKD releases to quay  OKD admins need permissions to annotate imagestreams and edit tags. This adds a new `okd-admins` group (myself, Clayton and Christian) and adds permissions to change imagestreams in `origin` namespace.

This is already applied in api.ci, but not yet in this git repo