oVirt fix upgrade jobs
On ovirt we noticed that ever since[1] was merged upgrade jobs started failing on the bootstrap phase of the first installation, meanining on installing the initial cluster.
This usually happens when there is a mismatch between the installer version and release image.
From looking at the logs I see that the release that is being installed is not the initial release, it is the latest.
This is due to the ENV var INSTALL_INITIAL_RELEASE which is not being set[2].
On the ovirt template we are checking for it[3].

We started porting to workflows now that we can acquire multiple leases on a job, but until that effort is done we want to upgrade jobs to run

[1] https://github.com/openshift/release/pull/14142
[2] https://github.com/openshift/release/pull/14142/files#diff-f20da9fe69faf7e3bfa2be8f95367c6ef26e92e246cad6c78424807b153cf18fL7298
[3] https://github.com/openshift/release/blob/master/ci-operator/templates/openshift/installer/cluster-launch-installer-ovirt-e2e.yaml#L384

Signed-off-by: Gal-Zaidman <gzaidman@redhat.com>