step-registry: don't use `base` as image
All steps that use the `base` image only need a few basic tools
available in the `centos:8` image. For steps that don't need specific
versions of an image based on what is version of openshift is being
tested, `from_image` should be used.  This also allows ocp nightly
builds (which don't include the `base` image) to be tested without having
to specify extra `base_images` in their ci-operator config.