Clean-up leased environments and detect broken state before running remote libvirt jobs.
The goal of this PR is to make no assumptions about the state of the environment lease when you get it. This allows us
to detect a bad state or connection issues to the environment before we start running the job. It will also
remove the need for manual clean-up of bad jobs since our job environments will be self-pruning. We know that the max
length of our infra names are 21 characters, so we use this knowledge to ensure we can inject the DNS records into
the right network.