image-puller rbacs: use rbac.authorization.k8s.io/v1
There is a difference between `rbac.authorization.k8s.io/v1` and
`authorization.openshift.io/v1`. The latter supports subjects of
`SystemGroup` kind (which seems legacy?) and the former does not, which
creates problems when people copy-paste content around.

The openshift resource seems to be converted to the matching upstream
resource on apply, so we can as well use the k8s one directly.