steps/openshift-e2e-test: Allow suite-conformance as a TEST_TYPE
The disruptive suites need to verify the cluster is in a sound state
afterwards. That action is best modelled today as "run custom suite, then
run suite conformance", similar to how upgrade then run conformance
apply. In the future this logic may move within openshift-tests but we
need to apply it across multiple versions and it is not clear yet what
use cases we will need.

This test type accepts TEST_SUITE and other standard flags.