release-controller: use interval instead of cron
For jobs that run on an interval, it is better to use the `interval`
field than `cron: @every` because `interval` is based on when the last
prowjobs were run. `cron: @every` instead runs the jobs immediately when
horologium starts, which can result in many jobs being suddenly run
whenever prow is updated and jobs being run more frequently than they
should.