ipi-aws: Use PVs for monitoring so that metrics survive upgrade
In the IPI AWS env we know that PV should be available dynamically,
so leverage that to enforce persistent volumes for metrics so that
metrics survive reschedules of pods. This is primarily for upgrade
and disruption jobs where we want to gain insight over the whole
run. This may become a more generic option for some job types in
the future or move to its own step that is shared for some job
types.