Adds openstack-install-ipi chain and step
The purpose of this step is to provide the developers with a test
they can execute on demand on submitted PRs.

The installation is executed with the following commands:

openstack-install-ip chain uses
openshift-installer create manifest
openshift-installer create ignition-configs
openshift-installer create cluster

All the commands are executed with TF_LOG=debug
and --log-level=debug by default for more detailed logs
but it may be modified via workflow environment definitions
(env variables).