SDN-1492: Network-tools: Change dockerfile path
The present Dockerfile file is named as "Dockerfile.rhel7" whereas it is
using the rhel8 builder. This is really confusing and in future would end
up in more confusion with each version.

This patch changes CI to use the default "Dockerfile" file in the repo so that the
file name doesn't cause confusion with the image type or version.

This was also a part of the suggestion of the security review for the repo.