step: GCP, AWS, and Azure clusters should get monitoring PVCs
All jobs on the main clouds should use a PVC for prometheus to
preserve data across upgrades and disruption. The specific reference
in the upgrade jobs is removed, and the bug in the previous step
corrected so that the right volume is referenced (older versions
may be regressed by this, but 4.7 and onwards keep working).