Tide: Allow 5k hourly tokens for sync
It appears the throttler does not work correctly, because unless the 5k
tokens are set as limit, it heavily throttles. Observing the actual
token usage on the v4 api, we seem to be using aroung 250 tokens/10
mins, which would be 1.5k/hour which is far from being problematic.

Should I be wrong and the etrapolation be too naive, the impact is tide
not working until the next token reset, which is IMHO an acceptable
risk.