Ensure o[-priv]/kubernetes prow conf can be automatically updated
Prow configuration for most repos is automatically updated when master
is frozen to require the `bugzilla/valid-bug` label to merge, and new
release branches automatically require `cherry-pick-approved` to
merge. Previously, configuration for o/kubernetes and o-priv/kubernetes
was maintained separately due to complications in maintaining the
owners file and this was preventing this automatic branch
configuration. Now that prow supports downstream owners, the
o/kubernetes repos can be included in the set of configuration that
will be automatically updated to ensure consistent process enforcement.