core-services/prow/02_config/_config: Drop redundant okd-machine-os Tide query
It's been part of a group that required only lgtm+approved on all
branches since 6a499ef6b4 (okd-machine-os: remove requirement for bug
and cherrypick labels, 2020-12-17, #14411).  But it was injected into
a query that required lgtm+approved for non-release branches in
f3db414a94 (make prow-config, 2021-02-01, #15412).  This commit
reverts that portion of the more recent commit, since the
non-release-branch query was redundant.