tide: enforce OCP merge criteria on o/cluster-cloud-controller-manager-operator
The repo was present in two of the three OCP merge criteria queries but
missing in the third, and at the same time it was present in the relaxed
critera query used for non-product repos, which essentially overrides
the OCP ones.

Fix the config so that OCP critera are enforced on the repo.