ci-tests: Fix approve csr function to not abort test on failure
The approve-csr function was created to approve csrs in the cluster
but in case of failure, the strategy is to keep trying until success.
With the current code, if the function fails, the test is aborted
due to the "set -euo pipefail" flag.

Adding a workaround that makes sure the oc command responsible for
approving the csr will always return success status as true with
a small OR true clause next to it. Since the while loop condition
is to check if files exist, this workaround will not mess with the
while condition.

Tested by creating a shell script with the set flag and running
a command with the or true clause, with and without a pipe near the
command.