WIP: ci-operator/step-registry/ipi/conf/workload: Synthetic workload step
Make it easier to turn up issues in CI that show up in the CI
clusters.  Those clusters are mostly full of CI jobs with moderate CPU
load and PodDisruptionBudgets that protect them from being evicted.
They run for up to 4 hours before being terminated, and have a 30
minute termination grace period on top of that.  We obviously can't
use workload that's that slow to drain in a CI job, or our CI job
would overshoot the limit and be killed.  In this commit, I'm adding a
new step (linked up just to the AWS update workflow for now) to
install a deployment that asks for 100m of CPU but then (I think)
consumes as much CPU as is available.  It would be awesome if there
was a test widget in some shipped container like tools that could be
configured to consume a particular amount of CPU and memory, although
I guess it would be hard to parameterize "regular" memory access.
Anyhow, this is a first-pass WIP to feel out this general approach.

The manifest will subsequently be picked up and fed to the installer
in the ipi-install-install step, or one of its close relatives.  We'd
be fine installing this as a day-2 manifest as well, but we don't have
tooling in place for that yet, and installing it via the installer
gives it more time to roll out into the compute nodes before the test
step rolls around.