OpenStack: more robust implementation of clouds.yaml to env vars
Previous implementation didn't account for multiple cloud entries in the
`clouds.yaml` causing it to generate bogus environment variables.

It would miss out a lot of the properties from `clouds.yaml`.

The new implementation relies on the fact that we can infer the environment
variable from the key name. According to openstacksdk documentation [1]:

  the keys in clouds.yaml correspond to the environment variable name
  without the `OS_` prefix and lowercase.

[1] https://docs.openstack.org/openstacksdk/latest/user/config/configuration.html