ci-operator: Update the image tags in the bundle stanza for master/4.8+ releases
Use the 4.8 image tags in the bundle substitution stanza for the master
and 4.8+ release branch configurations. Note: this was attempted before,
but these changes were only targeted to the release-4.8 branch
configuration (and not master), which resulted in the autoupdated from
the config-brancher to use the images that were specified in the master
branch configuration.