steps: Update cloud (AWS/GCP) to match Azure worker size of 4 core
The step registry lost the consistency injected into templates when
we introduced compact clusters (we set our default clusters for
compact to be 8 core x3, and our workers to 4 core x3). Restore that
so that all clouds have the same compute node size of 4 core 16gb,
which ensures that we are testing the median and recommended fleet
size.