OpenStack UPI: Remove unused functions
While looking to port ddbf50ab92c23e49043557a754b6cf9f4f96d446 to the
UPIt template, I noticed we didn't set the openstack environment
variables to run the cinder tests. This is not a big deal, we're going
to switch to the step-registry.

However, I took the opportunity to cleanup a bit the template and remove
the broken `get_clouds_param()` function. It was only used to set
`os_region_name` that was removed in
https://github.com/openshift/installer/commit/7a8be42aa8c0f690d59cd8acbc233f133009251c