Nuke jenkins-dev-operator
I was following a failure that started showing in the logs today, found
out the problem is related to the
`test_pull_request_online_registration` job on
openshift/online-registration whis does not work for some time and is
generally ignored on (rare) merges to the repo. In a Slack convo [2]
@dobbymoodge confirmed the job can be disabled, and because it was the
only job managed via the `jenkins-dev-operator`, the whole thing can go
down with it.

[1]
```json
{
   "component":"jenkins-operator",
   "error":"errors syncing: [failed to abort Jenkins build: response not 2XX: 400 Bad Request], errors reporting: []",
   "level":"error",
   "msg":"Error syncing.",
   "pod_name":"jenkins-dev-operator-688c4b4bc5-t6w4s",
   "pod_namespace":"ci",
   "severity":"error",
   "time":"2021-02-17T14:42:03Z"
}
```