coreos/rpm-ostree: Drop "build", wire up fcos-e2e test
This "fcos-e2e" test should run the code from https://github.com/coreos/rpm-ostree/pull/2593
Since this is doing a build, we can drop that CI context.
We can rely on `clang-build` to do the unit testing too (though
it is tempting to do the unit tests in the Dockerfile build too).