steps/openshift/e2e/test: Introduce the TEST_RUNS variable
Currently the openshift e2e workflows allow a job to skip some tests
by exporting the TEST_SKIPS variable, which value should be a valid POSIX Basic Regular
Expression (BRE). Now, suppose a job needs to run only a subset of the tests
(e.g. NodeConformance), but it cannot be expressed easily with a BRE the semantic of
"skip everything except for".

In order to solve that problem, this introduces the TEST_RUNS variable which likewise
accepts a BRE but rather indicate only the tests that should be ran (everything else
is skipped).

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>