csr: Add exit clause to approve_csr while loop
The approve_csr function in the e2e-gcp-upi tests does
not have an exit clause for when it is done with all
the approvals. This causes the function to run in
the background and causes the pods to not terminate once the
install is complete, causing it to timeout after 4 hours.

Added a condition that was in place in the templates file here in
steps in an attempt to end once all the approvals are done.