Refactor baremetalds-e2e-test to be more similar to openshift-e2e-test
This commits modifies baremetalds-e2e-test in order to align it, as much as possible (especially in terms of behaviors) to the openshift-e2e-test step,
and more in line with the other cluster profiles. This is also part of the ongoing effort to expand the OpenShift e2e CI coverage for the baremetal ipi platform,
so that it could be possible to apply the whole openshift/conformance suite.
Relevant changes introduced:
- The baremetal platform requires the tests image mirroring feature (see [1]) to execute some of the e2e tests.
-- Dev-scripts private registry has been used to store the mirrored test images.
-- Test images mirroring is currently available only in the openshift-tests binary for OpenShift 4.7 or greater. To maintain the backward compatibility, for jobs on branches < 4.7 a minimal list is adopted (defined by TEST_MINIMAL_LIST)
- RUN_UPGRADE_TEST has been deprecated
- baremetalds-e2e-conf step is not used anymore
- TEST_SKIPS is currently using to exclude temporarily those tests that are failing for which an analysis/fix is in progress