ci-operator/step-registry/ipi/install/vsphere: Drop install step
The ref was born in 6aa9f64f9f (Add VSphere IPI e2e steps, 2020-03-23, #7586)
so it could use the upi-installer image to work with a self-signed
certificate.  But vSphere installation hasn't required that since
5e8ba7c58e (vSphere: remove the use of upi-installer image; which will
prepare for upgrade testing, 2020-10-12, #12725).  Remove the
vSphere-specific step and point consumers at the generic step instead.

Same for the deprovision step.