Add periodic-ci-openshift-psap-ci-artifacts-release-*-gpu-operator-e2e* to testgrid
The GPU Operator is becoming a critical component for different
OpenShift projects, in particular for OpenDataHub. We need to be able
to highlight the stability of this Operator over the time, and the
TestGrid infrastructure looks like a good match to visualize Prow
nightly testing results over the time.

This PR includes the testing of the different versions of the GPU
Operator (1.4.0, 1.5.2, latest of OperatorHub and master branch
`release-next`), on OCP 4.5, 4.6 and 4.7.