Single node AWS tests limit parallel to 15
Duplicates the openshift/e2e/test step into single-node/e2e/test with
a small modification to change the number of parallel tests from 30
to 15. We've seen a lot of tests failing that we suspect are caused
by too much load on the single node. Hopefully this will make things
better.