Add timeout for Windows e2e test ref
This commit adds a custom timeout of 2h15m
to the Windows e2e test refs.
This change needs to be made to handle the
frequent 2h timeouts on e2e test job. This
change can be reverted in the future if the
tests from these jobs are split into individual
jobs and the tests complete within default
timeout.

commands ran:
make jobs and make validate-step-registry