ci-operator/config/openshift/origin/openshift-origin-master: Rollback AWS update jobs
Like the cluster-version operator has been doing since db0d358da9
(CVO: add rollback tests for 4.6+, 2020-10-14, #12789), and which we
used to do much more broadly until 666396269d (upgrade workflows:
remove abort-at from default TEST_OPTIONS, 2020-10-08, #12544).  I'm
personally in favor of the broad approach, because we want minor
rollbacks to work, so if we break master->PR rollback, we're in big
trouble.  But even with the narrow scoping for rollback jobs, origin's
test are implementing and monitoring the rollback functionality, so
they deserve to have a local test excercising it.

I've added this to the AWS update job, because Clayton wants it to be
optional, and the AWS update job has been optional and not always-run
since it was born in bd85872fe2 (add optional e2e-aws-upgrade job,
2020-08-10, #10785).