ci-operator/step-registry/openshift/e2e/test: Background and wait for openshift-tests
The end-to-end step originated without the background/wait pattern in
a2fd8c3bfb (step-registry: add Origin E2E test step, 2020-01-31, #6965).
But as described in 4472ace120
(ci-operator/templates/openshift/installer: Restore backgrounded
'create cluster', 2019-01-23, #2680), we want the background-and-wait
pattern to make openshift-tests a child process that will receive
TERMs via this step's existing 'trap' handler.  openshift-tests can
take hours, and when the step gets a TERM, we want to quickly pass
that through to the slow children, so they exit, so the step can exit,
so we have plenty of time for asset uploading and subsequent gather
and teardown steps.  That gives us the resources we need to figure out
why the test was abnormally slow.