ceph-csi: add automation to build image
The tests are using tools from Fedora, and the binary that is compiled
links against Ceph libraries. This makes the executable different from
the version that is released, based on the latest Ceph (CentOS) images.

Therefor, it is needed to build the final executable against different
libraries, on an other OS. Ideally the `make build` step in the tests
compiles and links the executable on a ceph/daemon-base:latest. Until
that is the case, we need a liteal Dockerfile doing the build again.