aws: add openshift-e2e-aws-byohostedzone workflow
Add the openshift-e2e-aws-byohostedzone workflow. The new workflow
behaves similarly to the openshift-e2e-aws-sharednetwork workflow.
Additionally, the new workflow creates a private hostedzone attached
to the VPC created for the shared network. This private hostedzone
is supplied to the installer for use by the cluster rather than
having the installer create its own hostedzone.

https://issues.redhat.com/browse/CORS-1667