supplemental images: stop building on config change
Unfortunately it does not seem that a ConfigChange trigger actually
triggers only on a real change to config; instead the trigger happens
whenever the object has it's generation bumped whatsoever. We want to
cache things that depend on this build, and we can't do that when the
root changes every couple of minutes during a working day as the
BuildConfig is re-applied to the cluster on postsubmit.

Signed-off-by: Steve Kuznetsov <skuznets@redhat.com>