Add debug machineSet and MCP
Add a new debug machineSet and MCP and create
a MC that changes the nvme.write_queues kernel
arg value to 1.
This will help verify the iostat data we are seeing
where reads are being starved when a lot of writes
are happening, hence setting up a limit to the write
queues.

Signed-off-by: Urvashi Mohnani <umohnani@redhat.com>