ci-operator: Avoid marking the any of the images[*] as optional
Update the openshift/operator-framework-olm prow configuration and avoid
marking any of the images listed in the images stanza as optional. This
was previously added as a way to get some CI feedback on whether the
images were able to be built successfully, but not blocking PRs from
merging due to failures.

That assumption is not true, and when specifying `optional: true` for
any images defined in that list of images, that image won't be built or
promoted externally.