baremetal: fix test matches and move 4.7 to minimal
A variety of patches to origin need to be backported to 4.7 to let us
expand the tests there. For now, we'll only use 4.8+ for the list of
skips so we can work on making it shorter. There's at least 10 PR's in
flight to fix various tests that need backporting from kubernetes,
origin, etc, so it's not feasible to deal with 4.7 and 4.8
concurrently.

This change also removes the tags on the end of test names, as the
annotations are changing with origin getting support for skipping
certain tests on Disconnected.  Leaving the annotations on the end of
our skip list causes them to no longer be skipped when annnotations are
added.