vsphere: Fallback to pip2 when there is no pip3
The vsphere DNS configuration fails if there is no pip3 installed
but there are CI jobs that run on pip2 for example the periodic
nightly 4.5 e2e vsphere tests that keep failing due to unavailable
pip3.

Adding a fallback to use pip2 in case pip3 was not available in
order for the CI jobs to pass.

Related bug: https://bugzilla.redhat.com/show_bug.cgi?id=1936867