kata-containers: Use CentOS 8 build root
Recently Kata Containers updated to Kernel 5.10 which requires a newer gcc version
than the one provided by the golang-1.13 (CentOS 7 based) image.

On kata-containers repository was added the Dockerfile which is used to build a build root
image based on CentOS 8 (newer gcc). So this changes the operator to use that Dockerfile.

Signed-off-by: Wainer dos Santos Moschetta <wainersm@redhat.com>