boilerplate image-v0.5.0=>image-v0.5.1
Boilerplate image-v0.5.0 was still trying to use
registry.svc.ci.openshift.org, but that was decommissioned last week, so
that image will never build.
https://github.com/openshift/boilerplate/pull/125 publishes image-v0.5.1
against the proper backing image. Update the ImageStream and DPTP-1640
workaround accordingly.

NOTE: This is unusual in that normally such changes are additive. But
since image-v0.5.0 does not and never will exist, we can and should get
rid of references to it.